+++ org-slipbox.el
@@ -62,6 +62,8 @@
 ;;;###autoload(autoload 'org-slipbox-node-random "org-slipbox-node" nil t)
 ;;;###autoload(autoload 'org-slipbox-node-insert "org-slipbox-node" nil t)
 ;;;###autoload(autoload 'org-slipbox-node-backlinks "org-slipbox-node" nil t)
+;;;###autoload(autoload 'org-slipbox-demote-entire-buffer "org-slipbox-edit" nil t)
+;;;###autoload(autoload 'org-slipbox-promote-entire-buffer "org-slipbox-edit" nil t)
 ;;;###autoload(autoload 'org-slipbox-refile "org-slipbox-edit" nil t)
 ;;;###autoload(autoload 'org-slipbox-extract-subtree "org-slipbox-edit" nil t)
 ;;;###autoload(autoload 'org-slipbox-link-replace-all "org-slipbox-link" nil t)